change: updatePrefuctureBlockを切り出し
