+++ src/js/modules/selectBlock.js
@@ -5,7 +5,7 @@
  * @param {string} dataValue
  * @returns {void} 返り値なし
  */
-export const deleteAreaBlock = (dataKey, dataValue) => {
+const deleteAreaBlock = (dataKey, dataValue) => {
   /** @type {HTMLMenuElement | null} ボタンリストのElement */
   const blockElement = document.querySelector(`[${dataKey}="${dataValue}"]`);
   if (!(blockElement instanceof HTMLMenuElement)) return;
@@ -46,4 +46,72 @@ export const createSelectBlock = (areaData, dataKey) => {
   });
 
   listElement.append(fragment);
+};
+
+/**
+ * 都道府県ボタンのElementを作成する関数
+ * @function
+ * @param {{key: string, description: string, list: Array<string>}} regionData
+ * @param {{key: string, description: string, prefectureList: Array<{name: string, ja:string, region: string}>}} prefectureRowData
+ * @param {string} dataKey
+ * @returns {void} 返り値なし
+ */
+export const updatePrefectureBlock = (
+  regionData,
+  prefectureRowData,
+  dataKey,
+) => {
+  const { key: regionKey } = regionData;
+  const { key: prefectureKey, prefectureList } = prefectureRowData;
+  /** @type {NodeListOf<HTMLInputElement>} */
+  const regionInputs = document.getElementsByName(regionKey);
+
+  /**
+   * 地域に合致する都道府県の一覧を返す関数
+   * @function
+   * @param {string} regionName
+   * @param {Array<{name: string, ja:string, region: string}>} prefectureList
+   * @returns {Array<string>}
+   */
+  const getMatchList = (regionName, prefectureList) => {
+    const matchList = prefectureList
+      .filter((prefecture) => prefecture.region === regionName)
+      .map((item) => item.ja);
+    return matchList;
+  };
+
+  /**
+   * 都道県データを整形する関数
+   * @function
+   * @param {{key: string, description: string, prefectureList: Array<{name: string, ja:string, region: string}>}} prefectureRowData
+   * @param {Array<string>} matchList
+   * @returns {{key: string, description: string, list: Array<string>}}
+   */
+  const formatPrefectureData = (prefectureRowData, matchList) => {
+    const { key, description } = prefectureRowData;
+    const formatData = {
+      key: key,
+      description: description,
+      list: matchList,
+    };
+    return formatData;
+  };
+
+  regionInputs?.forEach((input) => {
+    input.addEventListener("change", (event) => {
+      const targetInput = event.target;
+      if (!(targetInput instanceof EventTarget)) return;
+      /** @type {string} */
+      const regionName = targetInput.value;
+      if (regionName === "") return;
+
+      const prefectureNameList = getMatchList(regionName, prefectureList);
+      const prefectureData = formatPrefectureData(
+        prefectureRowData,
+        prefectureNameList,
+      );
+      deleteAreaBlock("data-weather-list", prefectureKey);
+      createSelectBlock(prefectureData, dataKey);
+    });
+  });
 };
